+++ python/CHANGES.rst
@@ -12,6 +12,7 @@ Changelog for tchannel.py
 
 - Allow custom handlers for unrecognized endpoints.
 - Released `tchannel.sync.TChannelSyncClient` and `tchannel.sync.thrift.client_for`.
+- Support exponential backoff in tchannel advertise method.
 
 
 0.8.5 (2015-06-30)